+++ examples/tools/inject-block/files/aliases.sh
@@ -1,0 +1,2 @@
+alias ll='ls -lah'
+alias gs='git status --short'